+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 1.4.7
 * Allow commands from any channel.
-* Start tournament if it's pending.
+* Start tournament from pending state.
 
 ### 1.4.6
 * Fix bug in ending tournaments.